fix: move reset to `<AppContext />`
