+++ src/components/MultiHopTrade/hooks/useAccountIds.tsx
@@ -1,10 +1,6 @@
 import type { AccountId } from '@shapeshiftoss/caip'
-import { useCallback, useEffect, useMemo } from 'react'
-import {
-  selectFirstHopSellAccountId,
-  selectLastHopBuyAccountId,
-  selectWalletId,
-} from 'state/slices/selectors'
+import { useCallback, useMemo } from 'react'
+import { selectFirstHopSellAccountId, selectLastHopBuyAccountId } from 'state/slices/selectors'
 import { tradeInput } from 'state/slices/tradeInputSlice/tradeInputSlice'
 import { useAppDispatch, useAppSelector } from 'state/store'
 
@@ -16,16 +12,9 @@ export const useAccountIds = (): {
 } => {
   const dispatch = useAppDispatch()
 
-  const walletId = useAppSelector(selectWalletId)
   // Default sellAssetAccountId selection
   const sellAssetAccountId = useAppSelector(selectFirstHopSellAccountId)
   const buyAssetAccountId = useAppSelector(selectLastHopBuyAccountId)
-
-  // Resets the sell and buy asset AccountIDs on wallet change to that we don't get stale trade input account selections while we're loading the new wallet
-  useEffect(() => {
-    dispatch(tradeInput.actions.setSellAssetAccountId(undefined))
-    dispatch(tradeInput.actions.setBuyAssetAccountId(undefined))
-  }, [dispatch, walletId])
 
   // Setters - the selectors above initially select a *default* value, but eventually onAccountIdChange may fire if the user changes the account
